Added final implementation

+++ js/GPACalc.js
@@ -46,7 +46,7 @@ function newCourse(semID) {
 }
 
 function deleteSemester(id) {
-    $("div[name=semester-"+id+"]").remove();;
+    $("div[name=semester-"+id+"]").remove();
     removeA(activeSemesterIDs, parseFloat(id));
     activeCourseIDs[id]=[];
     if(activeSemesterIDs.length < 1) {
@@ -67,15 +67,15 @@ function deleteCourse(semID,cID) {
 
 function printNewSemesterToScreen() {
     return '<div class="semester" name="semester-' + uniqueSemesterID + '">' + 
-        '<div class="semesterName">Semester Name: ' +
-            '<input type="text" name="semesterName-' + uniqueSemesterID + '">' +
-            '<button type="submit" class="deleteSemester deletebtn" name="'+ uniqueSemesterID +'">Delete</button>' +
+        '<div>' +
+            '<input type="text" class="semesterName" name="semesterName-' + uniqueSemesterID + '" placeholder = "Semester Name">' +
+            '<input class="deletebtn deleteSemester" name="'+ uniqueSemesterID +'" type="image" src="media/delete.png">' +
         '</div>' +
     '<div>' +
-      '<table class="classList-'+ uniqueSemesterID +'">' +
+      '<table class="classList classList-'+ uniqueSemesterID +'">' +
         '<!--Heading for the course list table for this semester-->' +
         '<tr>'+
-            '<th>Major?</th><th>Course Name</th><th>Grade</th><th>Credit #</th><th>GPA</th><th></th>'+
+            '<th>Major?</th><th>Course Name</th><th>Grade</th><th>Credit</th><th>GPA</th><th></th>'+
         '</tr>' +
         '<!--This will be a div that gets duplicated and removed as classes get added or removed dynamically-->' +
         '<!--Here is where other courses would get added -->' +
@@ -85,7 +85,7 @@ function printNewSemesterToScreen() {
         '<div class="leftGPA">GPA: <div name="GPA-' + uniqueSemesterID + '">N/a</div></div>' + 
         '<div class="rightGPA">Major GPA: <div name="majorGPA-' + uniqueSemesterID + '">N/a</div></div>' +
         '<!--button to add course to this semester-->' +
-        '<button class="button addCourse" name="' + uniqueSemesterID + '">Add course</button>' +
+        '<button class="button addCourse" name="' + uniqueSemesterID + '"><img src="media/add.png"><p>Add course</p></button>' +
       '</div>' +
     '</div>' +
   '</div>';
@@ -94,11 +94,28 @@ function printNewSemesterToScreen() {
 function printNewCourseToScreen(semesterID) {
     return newDiv = '<tr name="course-'+ semesterID +'-'+ uniqueCourseID +'">'+
       '<td><input type="checkbox" name="major-' + semesterID + '-' + uniqueCourseID + '" value="major"></td>' +
-      '<td><input type="text" name="courseName-' + semesterID + '-' + uniqueCourseID + '"></td>' + 
-      '<td><input type="text" name="courseGrade-' + semesterID + '-' + uniqueCourseID + '"></td>' +
-      '<td><input type="text" name="creditNumber-' + semesterID + '-' + uniqueCourseID + '"></td>' +
+      '<td><input type="text" name="courseName-' + semesterID + '-' + uniqueCourseID + '" size="35" placeholder="Course Name"></td>' + 
+      '<td><select name="courseGrade-' + semesterID + '-' + uniqueCourseID + '">'+
+        '<option value="A">A</option>'+
+        '<option value="A-">A-</option>'+
+        '<option value="B+">B+</option>'+
+        '<option value="B">B</option>'+
+        '<option value="B-">B-</option>'+
+        '<option value="C+">C+</option>'+
+        '<option value="C">C</option>'+
+        '<option value="C-">C-</option>'+
+        '<option value="D+">D+</option>'+
+        '<option value="D">D</option>'+
+        '<option value="F">F</option>'+
+        '</select></td>' +
+      '<td><select name="creditNumber-' + semesterID + '-' + uniqueCourseID + '">'+
+      '<option value="1">1 Credits</option>'+
+      '<option value="2">2 Credits</option>'+
+      '<option value="3">3 Credits</option>'+
+      '<option value="4">4 Credits</option>'+
+        '</select></td>' +
       '<td><div name="GPAOutput-' + semesterID + '-' + uniqueCourseID + '">N/a</div></td>' +
-      '<td><button  class="deletebtn deleteCourse" type="submit" semester="'+ semesterID +'" course="' + uniqueCourseID + '">Delete</button></td>' +
+      '<td><input class="deletebtn deleteCourse" semester="'+ semesterID +'"  course="' + uniqueCourseID + '" type="image" src="media/delete.png"></td>' +
     '</tr>';
 }
 
@@ -107,7 +124,7 @@ function saveToJson() {
         {"GPA":"" ,"MajorGPA":"" ,"semesters":[]};
     var GPA = $('div[name=GPA]').text();
     var majorGPA = $('div[name=majorGPA]').text();
-	
+
 	var message = "Data Saved: ";
 
     outJ.GPA = GPA; 
@@ -120,7 +137,7 @@ function saveToJson() {
 
         var sem = activeSemesterIDs[s];
         var semesterName = $('input[name=semesterName-'+sem+']').val();
-		
+
         //Check semester name
         if(semesterName != ""){
             outJ.semesters[s].name = semesterName;
@@ -134,14 +151,15 @@ function saveToJson() {
         majorGPA = $('div[name=majorGPA-'+sem+']').text();
 
         outJ.semesters[s].GPA = GPA; 
-        outJ.semesters[s].MajorGPA = majorGPA; 
-		
+        outJ.semesters[s].MajorGPA =majorGPA; 
+        
         for( var c = 0; c < activeCourseIDs[sem].length; c++) {
+            if($('input[name=courseName-'+sem+'-'+cor+']').val() != "") {
             outJ.semesters[s].courses.push(
                 { "name":"" ,"major":"" ,"credits":"" ,"grade":"" ,"GPA":""}
             )
             var cor = activeCourseIDs[sem][c];
-			
+
             //get value of major course checkbox
             var majorCourse = false;
             if($('input[name=major-'+sem+'-'+cor+']').is(':checked')) {
@@ -149,13 +167,13 @@ function saveToJson() {
             }
             //get value of course Name
             var courseName = $('input[name=courseName-'+sem+'-'+cor+']').val();
-            //get value of credit Number
-            var creditNumber = parseFloat($('input[name=creditNumber-'+sem+'-'+cor+']').val());
+            //get value of credit Number $('#Crd option:selected').text();
+            var creditNumber = parseFloat($('select[name=creditNumber-'+sem+'-'+cor+'] option:selected').text());
             //get value of grade
-            var courseGrade = $('input[name=courseGrade-'+sem+'-'+cor+']').val().toUpperCase();
+            var courseGrade = $('select[name=courseGrade-'+sem+'-'+cor+'] option:selected').text();
             //getValue of GPA
             var GPA = $('div[name=GPAOutput-'+sem+'-'+cor+']').text();
-			
+
             // Check Courses
             if(GPA != "N/a") {
                 if(courseName != ""){
@@ -165,24 +183,25 @@ function saveToJson() {
                     outJ.semesters[s].courses[c].name = "course-" + (c+1); 
                 }
 
-                outJ.semesters[s].courses[c].major = majorCourse;
-                outJ.semesters[s].courses[c].credits = creditNumber;
-                outJ.semesters[s].courses[c].grade = courseGrade;
-                outJ.semesters[s].courses[c].GPA = GPA;
+                outJ.semesters[s].courses[c].major = majorCourse
+                outJ.semesters[s].courses[c].credits = creditNumber
+                outJ.semesters[s].courses[c].grade = courseGrade
+                outJ.semesters[s].courses[c].GPA = GPA
             }
 
         }// For - courses
+    }
     }// For - Semester
-
+    
     // TODO: remove this alert once default values are added
-    alert(message); // Show any missing info that has been auto filled.
-	
-    if(sessionStorage.getItem("json")==null)
-    {
-        userData.semesterData = outJ;
-        sessionStorage.setItem('json', JSON.stringify(userData));
+    //alert(message); // Show any missing info that has been auto filled.
+    console.log(outJ);
+    if(userData == null) { //no account yet redirect with the option to use this data on a new account
+        sessionStorage.setItem('newjson', JSON.stringify(outJ));
     } else {
-        sessionStorage.setItem('newUserSemesterJson', JSON.stringify(outJ));
+    userData.semesterData = outJ;
+    sessionStorage.setItem('json', JSON.stringify(userData));
+        //sessionStorage.setItem('newUserSemesterJson', JSON.stringify(outJ));
     }
     
 }
@@ -198,68 +217,66 @@ function updateScreen() {
     var runningOMGPA = 0; // overall major gpa
 
     for(var s = 0; s < activeSemesterIDs.length; s++) {
-        console.log("test");
         var sem = activeSemesterIDs[s];
-		
+
 		var semName = $('input[name=semesterName-'+sem+']').val();
         if(semName == ""){
             semName = "semester-" + (s+1); 
         }
-		
+
         runningCN = 0;
         runningGPA = 0;
         runningMCN = 0;
         runningMGPA = 0;
-		
-            console.log("test:"+activeCourseIDs[sem]);
-            for( var c = 0; c < activeCourseIDs[sem].length; c++ ) {
+
+            console.log("test:"+activeCourseIDs[sem])
+            for( var c = 0; c < activeCourseIDs[sem].length; c++) {
                 var cor = activeCourseIDs[sem][c];
-				
+
 				var corName = $('input[name=courseName-'+sem+'-'+cor+']').val();
                 if(corName == ""){
                     corName = "course-" + (c+1); 
                 }
-				
-                var grade = $('input[name=courseGrade-'+sem+'-'+cor+']').val();
+
+                var grade = $('select[name=courseGrade-'+sem+'-'+cor+'] option:selected').text();
                 grade = grade.toUpperCase();
-                var gpa = "N/a";
-				
+                var gpa = "N/a"
+            
                 if(grade == 'A') {
-                    gpa = "4.00";
+                    gpa = "4.00"
                 } else if(grade == 'A-') {
-                    gpa = "3.70";
+                    gpa = "3.70"
                 } else if(grade == 'B+') {
-                    gpa = "3.30";
+                    gpa = "3.30"
                 } else if(grade == 'B') {
-                    gpa = "3.00";
+                    gpa = "3.00"
                 } else if(grade == 'B-') {
-                    gpa = "2.70";
+                    gpa = "2.70"
                 } else if(grade == 'C+') {
-                    gpa = "2.30";
+                    gpa = "2.30"
                 } else if(grade == 'C') {
-                    gpa = "2.00";
+                    gpa = "2.00"
                 } else if(grade == 'C-') {
-                    gpa = "1.70";
+                    gpa = "1.70"
                 } else if(grade == 'D+') {
-                    gpa = "1.30";
+                    gpa = "1.30"
                 } else if(grade == 'D') {
-                    gpa = "1.00";
+                    gpa = "1.00"
                 } else if(grade == 'F') {
-                    gpa = "0.00";
+                    gpa = "0.00"
                 } else if( grade == "" ) {
 					//NOP
-				}
-				else { // Invalid input
+				} else { // Invalid input
                    alert("Error: Grade is not Valid. " + semName + " - " + corName +"'s GPA Cannot be calculated - ");
                 }
 
-                var creditNumber = parseInt($('input[name=creditNumber-'+sem+'-'+cor+']').val());
-
+                var creditNumber = parseFloat($('select[name=creditNumber-'+sem+'-'+cor+'] option:selected').text());
+                
                 if( gpa != "N/a" ) // Don't bother calculating if grade is bad
                 {
                     if( isNaN(creditNumber))
                     {
-                        alert("Error: Credit amount must be a number. " + semName + " - " + corName +"'s GPA Cannot be calculated");
+                        //alert("Error: Credit amount must be a number. " + semName + " - " + corName +"'s GPA Cannot be calculated");
                         gpa = "N/a";
                     }
                     else if( (creditNumber < 1 || creditNumber > 4) )
@@ -279,33 +296,33 @@ function updateScreen() {
                     }
                 }
 
-                $('div[name=GPAOutput-'+sem+'-'+cor+']').html(gpa);
+                $('div[name=GPAOutput-'+sem+'-'+cor+']').html(gpa)
             } // For Courses
-			
+
         var finalGPA = (runningGPA/runningCN).toFixed(2);
         var finalMGPA = (runningMGPA/runningMCN).toFixed(2);
         if(isNaN(finalGPA)) {
-            finalGPA = "N/a";
+            finalGPA = "N/a"
         }
         if(isNaN(finalMGPA)) {
-            finalMGPA = "N/a";
+            finalMGPA = "N/a"
         }
-        $('div[name=GPA-'+sem+']').html(finalGPA);
-        $('div[name=majorGPA-'+sem+']').html(finalMGPA);
+        $('div[name=GPA-'+sem+']').html(finalGPA)
+        $('div[name=majorGPA-'+sem+']').html(finalMGPA)
         runningOCN = runningOCN + runningCN; // overall credit num
         runningOGPA = runningOGPA + runningGPA; // overall gpa
         runningOMCN = runningOMCN + runningMCN; // overall major credit num
         runningOMGPA = runningOMGPA + runningMGPA; // overall major gpa
     } // For semesters
-	
+
     var finalOGPA = (runningOGPA/runningOCN).toFixed(2);
     var finalMOGPA = (runningOMGPA/runningOMCN).toFixed(2);
     if(isNaN(finalOGPA)) {
-        finalOGPA = "N/a";
+        finalOGPA = "N/a"
     }
 
     if(isNaN(finalMOGPA)) {
-        finalMOGPA = "N/a";
+        finalMOGPA = "N/a"
     }
 
     $('div[name=GPA]').html(finalOGPA)
@@ -319,22 +336,22 @@ function loadJson(data) {
     activeCourseIDs = [[]];
 
     $('div[class=semesterList]').html("");
-	
-	var tUCID = 0;
-	
+
+    var tUCID = 0
+
     for(var s = 0; s < data.semesters.length; s++) {
 
         newSemester(data.semesters[s].courses.length);
         $("input[name=semesterName-"+s+"]").val(data.semesters[s].name);
 
         for( var c = 0; c < data.semesters[s].courses.length; c++) {
-            if(data.semesters[s].courses[c].major == 'true') {
+            if(data.semesters[s].courses[c].major == true) {
                 $('input[name=major-'+s+'-'+tUCID+']').attr('checked',true);
             }
-            $('input[name=major-'+s+'-'+tUCID+']').attr('checked');
+            //$('input[name=major-'+s+'-'+tUCID+']').attr('checked');
             $('input[name=courseName-'+s+'-'+tUCID+']').val(data.semesters[s].courses[c].name);
-            $('input[name=creditNumber-'+s+'-'+tUCID+']').val(data.semesters[s].courses[c].credits);
-            $('input[name=courseGrade-'+s+'-'+tUCID+']').val(data.semesters[s].courses[c].grade);
+            $('select[name=creditNumber-'+s+'-'+tUCID+']').val(data.semesters[s].courses[c].credits);
+            $('select[name=courseGrade-'+s+'-'+tUCID+']').val(data.semesters[s].courses[c].grade);
             tUCID++;
         }
     }
@@ -345,14 +362,13 @@ function loadJson(data) {
 $(document).ready(function () {
     userData = JSON.parse(sessionStorage.getItem('json'));
     console.log(userData);
-    if(userData != null)
+    if(userData == null || userData.semesterData.semesters == "")
     {
-        loadJson(userData.semesterData)
-    } else {
         newSemester(4);
+    } else {
+        loadJson(userData.semesterData)
     }
 });
-
 
 $('button[name=addSemester]').on('click',function() { //add new semester
     newSemester(4);
@@ -363,8 +379,13 @@ $('.semesterList').on('click','.addCourse',function() {
 });
 
 $('.semesterList').on('click','.deleteSemester',function() {
-    deleteSemester($(this).attr('name'));
-    updateScreen();
+
+    if( confirm("Are you sure you want to delete this semester?") )
+    {
+        deleteSemester($(this).attr('name'));
+        updateScreen();
+    }
+    
 });
 
 $('.semesterList').on('click','.deleteCourse',function() {
